修改GetSysClockTime返回值

+++ frameworks/proxy/events/src/manipulation_event.cpp
@@ -25,9 +25,9 @@ namespace OHOS {
     }
 
 ManipulationEvent::~ManipulationEvent() {};
-void ManipulationEvent::Initialize(int32_t windowId, int32_t startTime, int32_t operationState, int32_t pointerCount,
+void ManipulationEvent::Initialize(int32_t windowId, int64_t startTime, int32_t operationState, int32_t pointerCount,
                                    fingerInfos fingersInfos[], int32_t highLevelEvent, const std::string& uuid,
-                                   int32_t sourceType, uint64_t occurredTime, const std::string& deviceId,
+                                   int32_t sourceType, int64_t occurredTime, const std::string& deviceId,
                                    int32_t inputDeviceId,  bool isHighLevelEvent, uint16_t deviceUdevTags)
 {
     if (pointerCount < 0 || pointerCount > FINGER_NUM) {
@@ -61,7 +61,7 @@ void ManipulationEvent::Initialize(ManipulationEvent& maniPulationEvent)
     CHK(ret == EOK, MEMCPY_SEC_FUN_FAIL);
 }
 
-int32_t ManipulationEvent::GetStartTime() const
+int64_t ManipulationEvent::GetStartTime() const
 {
     return startTime_;
 }
