+++ interfaces/native/innerkits/event/include/speech_event.h
@@ -114,7 +114,7 @@ public:
     * @since 1
     */
     void Initialize(int32_t windowId, int32_t action, int32_t scene, int32_t mode, const std::string& actionProperty,
-                    int32_t highLevelEvent, const std::string& uuid, int32_t sourceType, int32_t occurredTime,
+                    int32_t highLevelEvent, const std::string& uuid, int32_t sourceType, int64_t occurredTime,
                     const std::string& deviceId, int32_t inputDeviceId, bool isHighLevelEvent,
                     uint16_t deviceUdevTags = 0);
